requirements.txt
Django==3.2
djangorestframework==3.13.1
django-cors-headers==3.10.1
channels==3.0.4
daphne==3.0.2
django-heroku
psycopg2
dj-database-url
whitenoise
